inital routes and models moved to correct file path
