+++ .env
@@ -1,0 +1,3 @@
+DB_NAME='kursties_kookies_db'
+DB_USER='root'
+DB_PASSWORD=''
