+++ models/Product.js
@@ -30,13 +30,13 @@ Product.init(
 				isNumeric: true,
 			},
 		},
-		category_id: {
-			type: DataTypes.INTEGER,
-			references: {
-				model: 'category',
-				key: 'id',
-			},
-		},
+		// category_id: {
+		// 	type: DataTypes.INTEGER,
+		// 	references: {
+		// 		model: 'category',
+		// 		key: 'id',
+		// 	},
+		// },
 	},
 	{
 		sequelize,
