+++ models/User.js
@@ -1,4 +1,4 @@
-const { Model, Datatypes } = require('sequelize');
+const { Model, DataTypes } = require('sequelize');
 const sequelize = require('../config/connection');
 const bcrypt = require('bcrypt');
 
@@ -13,17 +13,17 @@ class User extends Model {
 User.init(
 	{
 		id: {
-			type: Datatypes.INTEGER,
+			type: DataTypes.INTEGER,
 			allowNull: false,
 			primaryKey: true,
 			autoIncrement: true,
 		},
 		username: {
-			type: Datatypes.STRING,
+			type: DataTypes.STRING,
 			allowNull: false,
 		},
 		email: {
-			type: Datatypes.STRING,
+			type: DataTypes.STRING,
 			allowNull: false,
 			unique: true,
 			validate: {
@@ -31,20 +31,20 @@ User.init(
 			},
 		},
 		password: {
-			type: Datatypes.STRING,
+			type: DataTypes.STRING,
 			allowNull: false,
 			validate: {
 				len: [8],
 			},
 		},
-		cart: {
-			type: Datatypes.ARRAY,
-			allowNull: true,
-			references: {
-				model: 'cart',
-				key: 'id',
-			},
-		},
+		// cart: {
+		// 	type: DataTypes.ARRAY(DataTypes.STRING),
+		// 	allowNull: true,
+		// 	references: {
+		// 		model: 'cart',
+		// 		key: 'id',
+		// 	},
+		// },
 	},
 	{
 		hooks: {
